Refactor blockchain configuration and RPC URL handling
- Moved various blockchain constants and configurations to a centralized CHAIN_CONFIG object in the config.
- Removed redundant constants and functions related to blockchain chains in helper/index.js.
- Updated the RPC URL handling to utilize the new CHAIN_CONFIG structure for better maintainability.
- Cleaned up unused imports and constants in walletsSlice.js and rpcUrls.js.
